+++ src/common/gate.py
@@ -6,24 +6,24 @@ from common.express_lane import ExpressLane, express_lanes_by_start
 
 
 @dataclass
-class Checkpoint:
+class Gate:
     name: str
-    points: tuple[Coordinate, ...]
+    points: (Coordinate, Coordinate)
     express_lane_start: Coordinate
 
     @classmethod
-    def by_name(cls, name: str) -> "Checkpoint":
-        for checkpoint in checkpoints:
-            if checkpoint.name == name:
-                return checkpoint
+    def by_name(cls, name: str) -> "Gate":
+        for gate in gates:
+            if gate.name == name:
+                return gate
 
     @property
     def express_lane(self) -> ExpressLane:
         return express_lanes_by_start[self.express_lane_start]
 
 
-checkpoints = []
-with open("resources/checkpoints.csv") as f:
+gates = []
+with open("resources/gates.csv") as f:
     reader = csv.reader(f)
     next(reader, None)  # Skip header
     for row in reader:
@@ -32,5 +32,5 @@ with open("resources/checkpoints.csv") as f:
         express_lane_start = tuple(map(float, row[3:5]))
         next(reader, None)
         b = Coordinate(*map(float, row[1:3]))
-        checkpoint = Checkpoint(name, (a, b), Coordinate(*express_lane_start))
-        checkpoints.append(checkpoint)
+        gate = Gate(name, (a, b), Coordinate(*express_lane_start))
+        gates.append(gate)