+++ changelog.md
@@ -1,3 +1,4 @@
 # Changelog
 ## 2024-10-05
 - Added changelog file
+- Renamed `Checkpoint` to `Gate`